fix: 配置 Flutter 编译环境并修复 iOS 构建问题
- 替换 mediapipe_pose 为 google_mlkit_pose_detection
- 修复 pose_detector.dart 枚举字段命名冲突
- 修复 pose_processor.dart 适配 ML Kit API
- 修复 camera_preview.dart CameraLensDirection 导入
- 修复 realtime_pose_preview.dart 图像格式转换
- 修复 home_page.dart 导入冲突
- 修复 app_theme.dart CardThemeData 类型
- 修复 metrics_calculator.dart 返回类型不匹配
- 修复 scoring_engine_test.dart 逻辑错误
- 更新 .gitignore 添加 iOS/Android 编译临时文件
- 初始化 iOS/Android 平台配置
- 配置 iOS Podfile 支持 iOS 15.5+

编译测试: flutter analyze 通过, iOS 模拟器构建成功

+++ lib/analysis/metrics_calculator.dart
@@ -443,7 +443,7 @@ class SwingMetricsCalculator {
 
     // 返回时序差 (毫秒)
     // 正值表示髋部先启动
-    return shoulderStartTime - hipStartTime;
+    return (shoulderStartTime - hipStartTime).toDouble();
   }
 
   /// 获取转动变化量
